Move test to tests directory
